requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
python-multipart==0.0.9
jinja2==3.1.4
requests==2.32.3
python-dotenv==1.0.1
slowapi==0.1.9
aiofiles==23.2.1
